refactor: split contract into modular file structure
