+++ bindings/package.json
@@ -36,5 +36,6 @@
   },
   "devDependencies": {
     "typescript": "^5.6.2"
-  }
+  },
+  "packageManager": "yarn@4.7.0+sha512.5a0afa1d4c1d844b3447ee3319633797bcd6385d9a44be07993ae52ff4facabccafb4af5dcd1c2f9a94ac113e5e9ff56f6130431905884414229e284e37bb7c9"
 }
