+++ contracts/src/tests/mod.rs
@@ -1,0 +1,8 @@
+//! Test modules for the XLM Price Prediction Market contract.
+
+mod initialization;
+mod betting;
+mod resolution;
+mod lifecycle;
+mod edge_cases;
+
