+++ contracts/test_snapshots/tests/betting/test_get_user_position_no_bet.1.json
