+++ contracts/test_snapshots/tests/betting/test_place_bet_negative_amount.1.json
