+++ contracts/test_snapshots/tests/betting/test_place_bet_zero_amount.1.json
