+++ contracts/test_snapshots/tests/edge_cases/test_round_with_no_participants.1.json
