+++ contracts/test_snapshots/tests/initialization/test_balance_for_new_user.1.json
