+++ contracts/test_snapshots/tests/initialization/test_initialize.1.json
