+++ contracts/test_snapshots/tests/initialization/test_mint_initial.1.json
