+++ contracts/test_snapshots/tests/initialization/test_mint_initial_only_once.1.json
