+++ contracts/test_snapshots/tests/lifecycle/test_get_active_round_when_none.1.json
