+++ contracts/test_snapshots/tests/resolution/test_claim_winnings_when_none.1.json
