+++ contracts/test_snapshots/tests/resolution/test_resolve_round_price_went_down.1.json
