+++ contracts/test_snapshots/tests/resolution/test_resolve_round_price_went_up.1.json
